pyproject.toml
[tool.poetry]
name = "2022_mimiciv_vital_survival"
version = "0.1.0"
description = ""
authors = ["MinDong <mdskylover@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyMySQL = "^1.0.2"
pandas = "^1.4.0"
numpy = "^1.22.1"
python-dotenv = "^0.19.2"
tqdm = "^4.62.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.12b0"
ipykernel = "^6.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
